requirements.txt
# Backend
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.3
python-dotenv==1.0.1
aiohttp==3.13.5
pydantic==2.9.2

# ML / NLP
transformers==4.47.0
torch==2.6.0
scipy>=1.14.0
beautifulsoup4==4.12.3
langdetect==1.0.9

# Data APIs
google-api-python-client==2.133.0

# Utilities
tqdm==4.66.4
pandas>=2.3.0

nltk